move helper function to utils module. #88
